+++ springql/src/lib.rs
@@ -6,5 +6,8 @@
 
 pub use springql_core::api::*;
 
+mod config;
+pub use crate::config::config_from_toml;
+
 #[cfg(test)]
 mod tests {}